more progress
Took a break for a few days to work on other stuff. Worked on it a bit today and got the client and server parts talking to each other correctly. I think the threads all quit like they should although the close function callback is broken and wont ever work. Will have to just implement a periodic cleanup function instead.