More boredom = more updates
Reworked the initialization process of the OpenGL elements so that we can get parameters from the server if running in ClientMode. Now the server sends the window size and grid size to the clients which use that info when creating the window of their own. Also made handling of client thread termination a little better. Now to get them all talking to each other while playing the game....