Filling out client/server comms
Not a big update, just filled out some of the client/server communications. Largest change was allowing the server to change behavior depending on the type of reply.

There is a ton of work to be done rewriting the snake logic. Basically it all has to get gutted and most of it will get put into the server-side of things. Two steps forward and one step back is how this is going. I think I know what I am supposed to do and then I discover a few things I forgot about, making it all more complicated. Dis gonna be gud.